+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2025, Evan Low'
 author = 'Evan Low'
 
 # The full version, including alpha/beta/rc tags
-release = '0.13.0'
+release = '0.13.2'
 version = '0.13.0'